requirements.txt
numpy==1.26.4
pandas==2.1.4
streamlit==1.34.0
streamlit_paste_button==0.1.2
streamlit-image-select==0.6.0
streamlit-option-menu==0.3.12
Pillow==10.2.0
easyocr==1.7.1
opencv-python==4.9.0.80
ultralytics==8.2.2